Refactor app.js and apply vite
- Apply Vite for TDD, DX, e.t.c
- Divide app.js into multiple js files
- Remove app.js
- Add .gitignore for node_modules and dist
- Refactor some small code in index.html
